Remove u8 casts
This will work perfectly fine with integers

+++ src/synesthesia/common.nim
@@ -1,6 +1,6 @@
 type
   BFCore* = ref object
-   memory*: array[1024, uint8]
+   memory*: array[1024, int]
    pc*, ap*: int
 
   BFSymbolKind* = enum
